+++ documentation/ChangeLog.md
@@ -10,6 +10,7 @@
  * Fixed #2: Lifetimes are incorrectly hilighted as char
  * Fixed: "Attribute" syntax highlighting option missing from Source Coloring preference page.
  * Added: Can now do syntax highlighting for lifetime token.
+ * Added: Strings inside attribute now highlighted as well.
 
 ### RustDT 0.1.0 - Initial release 
  * Added: Rust source code editor, with: